Narrow SeqFn yaw to s16 before rotX divide
Target extsh's yaw before srawi in both behaviour-state turn arms.

+++ src/main/dll/dll_014C_babycloudrunner.c
@@ -966,7 +966,7 @@ int babycloudrunner_SeqFn(int* obj, int unused, ObjAnimUpdateState* animUpdate)
         animUpdate->hitVolumePair &= ~0x2;
         yaw = Obj_GetYawDeltaToObject((int)obj, player, 0);
         fn_8003ADC4(obj, (int*)player, (char*)sub + 0x3c, 0x28, 0, 3);
-        ((GameObject*)obj)->anim.rotX += yaw / 8;
+        ((GameObject*)obj)->anim.rotX += (s16)yaw / 8;
         if (inRange != 0)
         {
             animUpdateBytes[0x90] |= 4;
@@ -980,7 +980,7 @@ int babycloudrunner_SeqFn(int* obj, int unused, ObjAnimUpdateState* animUpdate)
         animUpdate->hitVolumePair &= ~0x2;
         yaw = Obj_GetYawDeltaToObject((int)obj, getTrickyObject(), 0);
         fn_8003ADC4(obj, getTrickyObject(), (char*)sub + 0x3c, 0x28, 0, 3);
-        ((GameObject*)obj)->anim.rotX += yaw / 8;
+        ((GameObject*)obj)->anim.rotX += (s16)yaw / 8;
         break;
     }
     return 0;
